Implement all states correctly

+++ server/src/main/java/org/cloudburstmc/server/item/ItemType.java
@@ -37,7 +37,7 @@ public interface ItemType {
         return 0;
     }
 
-    default int getFuelTime() {
+    default short getFuelTime() {
         return 0;
     }
